pyproject.toml
[project]
name = "agentic-workflow-builder"
version = "0.1.0"
description = ""
authors = [
    {name = "Pankaj Ganesh",email = "pankaj@sukshi.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13" 
dependencies = [
    "openai (>=1.66.3,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "langchain-openai (>=0.2.1,<0.3.0)",
    "langchain (>=0.3.21,<0.4.0)",
    "pymupdf (>=1.25.4,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "ollama (>=0.4.7,<0.5.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "mcp (>=1.8.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-core (>=0.3.64,<0.4.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "langtrace-python-sdk (>=3.8.21,<4.0.0)",
    "openlit (>=1.34.35,<2.0.0)",
    "opik (>=1.8.9,<2.0.0)",
    "streamlit (>=1.47.1,<2.0.0)",
    "clickhouse-connect (>=0.8.18,<0.9.0)",
    "graphviz (>=0.21,<0.22)",
    "langgraph (>=0.6.7,<0.7.0)",
    "langchain-mcp-adapters (>=0.1.9,<0.2.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "langchain-tavily (>=0.2.13,<0.3.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "google-cloud-storage (>=3.9.0,<4.0.0)",
    "cryptography (>=44.0.0,<45.0.0)",
    "qdrant-client (>=1.13.0,<2.0.0)",
]

[tool.poetry]
packages = []
package-mode = false

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
httpx = ">=0.27,<0.29"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
alembic = "^1.16.2"



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=./ --cov-report=term-missing"
asyncio_mode = "auto"
